refactor: Add a *consuming* data handle that removes a key (#4135)
This PR adds a consuming data handle, which removes a key from the whiteboard. I need this, because for the EDM4hep/PODIO writing, I need to take a `podio::Frame` that's stored in the whiteboard and move it into the PODIO IO system, and I don't want to copy the frame.
The handle is fully integrated with the dependency tracking system so it will trigger an error if a read handle tries to access a key that was removed.