refactor: Replace `SeedingPerformanceWriter` with `TrackFinderPerformanceWriter` in Examples (#4152)
`SeedingPerformanceWriter` and `TrackFinderPerformanceWriter` have a very similar purpose while one is implemented to strictly use seeds and the other to strictly use tracks. We already have infrastructure to convert seeds to prototracks and prototracks to tracks. This can be used to ship around `SeedingPerformanceWriter` and directly use `TrackFinderPerformanceWriter` which also supports fakes and has more plots.

To make this work the prototracks to tracks converter was enhanced to also use track parameters.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

- **New Features**
	- Integrated new track conversion and matching algorithms to enhance processing capabilities.
- **Refactor**
	- Improved track parameter handling with added checks for consistency.
- **Chores**
	- Removed outdated performance tracking components and unnecessary dependencies from build, tests, and scripting environments.
	- Eliminated references to the `SeedingPerformanceWriter` across various scripts and tests.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->